Modified predict.py - created generate_umap_plot function for plotting umap data.
